changed: put movie videodatabase methods in separate file
